# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CRUIS API - EBSCO Authentication and Proxy Service",
    "version": "1.0.0",
    "description": "Backend API server for EBSCO authentication and proxy services. This API provides endpoints to authenticate with EBSCO using library card credentials and proxy authenticated requests to EBSCO services."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/api/auth/ebsco": {
      "post": {
        "summary": "Authenticate with EBSCO",
        "description": "Authenticates with EBSCO using library card credentials and returns an authentication token.",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["cardNumber", "password"],
                "properties": {
                  "cardNumber": {
                    "type": "string",
                    "description": "Library card number",
                    "example": "1001600244772"
                  },
                  "password": {
                    "type": "string",
                    "description": "Library card password",
                    "example": "your-password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "authToken": {
                      "type": "string",
                      "description": "EBSCO authentication cookie value"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing cardNumber or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "message": {
            "type": "string",
            "description": "Detailed error message"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code"
          }
        }
      }
    }
  }
}